--- a/terraform/application/config/migration.tfvars.json
+++ b/terraform/application/config/migration.tfvars.json
@@ -8,7 +8,7 @@
     "redis_cache_sku_name": "Basic",
     "enable_logit": true,
     "webapp_replicas": 3,
-    "worker_replicas": 20,
+    "worker_replicas": 30,
     "webapp_memory_max": "4Gi",
     "worker_memory_max": "2Gi",
     "postgres_flexible_server_sku": "GP_Standard_D2ds_v5"