[Jira-3511](https://dfedigital.atlassian.net.mcas.ms/jira/software/projects/CPDLP/boards/87?assignee=712020%3A5b5839dc-e4d6-4df5-9dd6-627cc49fe345&selectedIssue=CPDLP-3511)

### Context

We want to utilise all available workers where possible by ensuring the records_per_worker split evenly across a multiple of the number of workers.

### Changes proposed in this pull request

- Optimise records_per_worker

Also increases to 30 workers for better throughput.

### Guidance to review

Main:

```
  [["application", "29 minutes and 25.89283599999999 seconds"],
 ["declaration", "4 minutes and 8.569072000000006 seconds"],
 ["statement_item", "2 minutes and 49.38589099999999 seconds"],
 ["user", "2 minutes and 21.038196 seconds"],
 ["contract", "1 minute and 59.136003 seconds"],
 ["application_state", "1 minute and 59.082032 seconds"],
 ["participant_outcome_api_request", "1 minute and 9.245193999999998 seconds"],
 ["participant_outcome", "1 minute and 0.7751540000000006 seconds"],
 ["participant_id_change", "35.009366 seconds"],
 ["school", "19.413319 seconds"],
 ["statement", "15.234018 seconds"],
 ["declaration_superseded_by", "8.076017 seconds"],
 ["private_childcare_provider", "6.151425 seconds"],
 ["itt_provider", "2.152112 seconds"],
 ["schedule", "1.187606 seconds"],
 ["course", "1.0787 seconds"],
 ["cohort", "1.055177 seconds"],
 ["lead_provider", "1.052633 seconds"]]
```

This branch: 
 
```
 [["application", "28 minutes and 36.58124500000008 seconds"],
 ["declaration", "3 minutes and 20.99262300000001 seconds"],
 ["user", "3 minutes and 1.441744 seconds"],
 ["statement_item", "2 minutes and 47.366584999999986 seconds"],
 ["application_state", "2 minutes and 17.27501000000001 seconds"],
 ["participant_id_change", "56.015721 seconds"],
 ["participant_outcome", "34.129365 seconds"],
 ["participant_outcome_api_request", "32.001896 seconds"],
 ["school", "13.180682 seconds"],
 ["contract", "12.058219 seconds"],
 ["private_childcare_provider", "5.90575 seconds"],
 ["declaration_superseded_by", "5.891588 seconds"],
 ["statement", "4.422493 seconds"],
 ["itt_provider", "4.343207 seconds"],
 ["schedule", "2.320284 seconds"],
 ["course", "1.262134 seconds"],
 ["lead_provider", "1.214325 seconds"],
 ["cohort", "1.206152 seconds"]]
```

| Main    | This Branch |
| -------- | ------- |
| <img width="1004" alt="main" src="https://github.com/user-attachments/assets/dcc8f90f-34bc-452b-9b53-253e4d339bd5">  | <img width="1007" alt="split-workers-better" src="https://github.com/user-attachments/assets/dec1534f-fb83-4abd-b253-5c3c2bcfb1cd">   |